+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-11-29
+
+### Added
+
+- `Emitter` implementation with custom emission strategy. This allows client code to determine
+*how* emitted events should trigger their listeners. Sequentially without caring about the result?
+Using `Promise.all`? Using `Promise.allSettled`? All supported.
+
+### Changed
+
+- Changed the `EmitterLikeBase` implementation to leverage the new `Emitter`. This is a breaking
+change, as the constructor now requires and argument and the class itself requires an extra
+generic parameter.
+
 ## [0.1.0] - 2025-11-08
 
 ### Added